requirements.txt
boto3
Flask
Flask-Caching
python-dotenv
requests
gunicorn
prometheus-flask-exporter
prometheus_client
